requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic[email]==2.11.5
scikit-learn==1.6.1
xgboost==3.0.0
pandas==2.2.3
python-dotenv==1.1.0
flask_sqlalchemy==3.0.5
sqlalchemy_filters==0.13.0
psycopg2-binary==2.9.10
matplotlib==3.10.1
keras==3.10.0
tensorflow==2.20.0
seaborn==0.13.2
scikeras==0.13.0
tqdm==4.67.1
python-multipart==0.0.20
pyodbc==5.2.0
alembic==1.16.1
pydantic-settings==2.9.1
python-jose[cryptography]==3.5.0
argon2-cffi==25.1.0
passlib==1.7.4
shap==0.44.1
numpy>=1.26.0,<2.0.0  # Compatible con Python 3.12 y SHAP